Deposits: connect deposit currency dropdown to deposit anchors (reorganized deposits tab logic)
